Lesson23 - Convert Core Data Model to Swift Data Model and convert code.
