--- Dex/Dex.xcdatamodeld/Dex.xcdatamodel/contents
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="23231" systemVersion="24E5228e" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="Pokemon" representedClassName="Pokemon" syncable="YES" codeGenerationType="class">
-        <attribute name="attack" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="defense" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="favorite" attributeType="Boolean" defaultValueString="NO" usesScalarValueType="YES"/>
-        <attribute name="hp" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="id" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="name" attributeType="String"/>
-        <attribute name="shiny" optional="YES" attributeType="Binary"/>
-        <attribute name="shinyURL" attributeType="URI"/>
-        <attribute name="specialAttack" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="specialDefense" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="speed" attributeType="Integer 16" defaultValueString="0" usesScalarValueType="YES"/>
-        <attribute name="sprite" optional="YES" attributeType="Binary"/>
-        <attribute name="spriteURL" attributeType="URI"/>
-        <attribute name="types" attributeType="Transformable" valueTransformerName="NSSecureUnarchiveFromData" customClassName="[String]"/>
-        <uniquenessConstraints>
-            <uniquenessConstraint>
-                <constraint value="id"/>
-            </uniquenessConstraint>
-        </uniquenessConstraints>
-    </entity>
-</model>
