--- Dex/FetchedPokemon.swift
@@ -1,100 +1,0 @@
-//
-//  FetchedPokemon.swift
-//  Dex
-//
-//  Created by Jim Rainville on 3/3/25.
-//
-
-import Foundation
-
-struct FetchedPokemon: Decodable {
-    let id: Int16
-    let name: String
-    let types: [String]
-    let hp: Int16
-    let attack: Int16
-    let defense: Int16
-    let specialAttack: Int16
-    let specialDefense: Int16
-    let speed: Int16
-    let spriteURL: URL
-    let shinyURL: URL
-    
-    // Coding keys are where you map the data you want to get out of the JSON to the actual JSON structure.
-    // This is the API call that gave us the sample data: https://pokeapi.co/api/v2/pokemon/1/
-    // Drop that into the json viewer here: https://jsonviewer.stack.hu
-    enum CodingKeys: String, CodingKey {
-        case id
-        case name
-        case types
-        case stats
-        case sprites
-        
-        enum TypeDictionaryKeys: CodingKey {
-            case type
-            enum TypeKeys: CodingKey {
-                case name
-            }
-        }
-        enum StatDictionaryKeys: CodingKey {
-            case baseStat
-        }
-        
-        enum SpriteKeys: String, CodingKey {
-            case spriteURL = "frontDefault"
-            case shinyURL = "frontShiny"
-        }
-        
-    }
-    
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        
-        id = try container.decode(Int16.self, forKey: .id)
-        name = try container.decode(String.self, forKey: .name)
-        var decodedTypes: [String] = []
-        // Keyed container is a dictionary, unkeyed containers are arrays.
-        var typesContainer = try container.nestedUnkeyedContainer(forKey: .types)
-        while !typesContainer.isAtEnd {
-            let typesDictionaryContainer = try typesContainer.nestedContainer(keyedBy: CodingKeys.TypeDictionaryKeys.self)
-            let typeContainer = try typesDictionaryContainer.nestedContainer(keyedBy: CodingKeys.TypeDictionaryKeys.TypeKeys.self, forKey: .type)
-            let type = try typeContainer.decode(String.self, forKey: .name)
-            decodedTypes.append(type)
-        }
-        
-        
-        if decodedTypes.count == 2 && decodedTypes[0] == "normal" {
-            /*
-            let tempType = decodedTypes[0]
-            decodedTypes[0] = decodedTypes[1]
-            decodedTypes[1] = tempType
-             */
-            // This is a shortcut for doing the above.
-            decodedTypes.swapAt(0, 1)
-        }
-        
-        
-        types = decodedTypes
-        
-        var decodedStats: [Int16] = []
-        var statsContainer = try container.nestedUnkeyedContainer(forKey: .stats)
-        while !statsContainer.isAtEnd {
-            let statsDictionaryContainer = try statsContainer.nestedContainer(keyedBy: CodingKeys.StatDictionaryKeys.self)
-            let stat = try statsDictionaryContainer.decode(Int16.self, forKey: .baseStat)
-            decodedStats.append(stat)
-        }
-        hp = decodedStats[0]
-        attack = decodedStats[1]
-        defense = decodedStats[2]
-        specialAttack = decodedStats[3]
-        specialDefense = decodedStats[4]
-        speed = decodedStats[5]
-        
-        
-        let spritesContainer = try container.nestedContainer(keyedBy: CodingKeys.SpriteKeys.self, forKey: .sprites)
-        spriteURL = try spritesContainer.decode(URL.self, forKey: .spriteURL)
-        shinyURL = try spritesContainer.decode(URL.self, forKey: .shinyURL)
-    }
-}
-
-
